+++ Tests/ConvertTo-JiraField.Tests.ps1
@@ -7,7 +7,7 @@ Describe "ConvertTo-JiraField" {
         . "$PSScriptRoot/Shared.ps1"
 
         $sampleJson = '{"id":"issuetype","name":"Issue Type","custom":false,"orderable":true,"navigable":true,"searchable":true,"clauseNames":["issuetype","type"],"schema":{"type":"issuetype","system":"issuetype"}}'
-        $sampleObject = ConvertFrom-Json2 -InputObject $sampleJson
+        $sampleObject = ConvertFrom-Json -InputObject $sampleJson
 
         $r = ConvertTo-JiraField $sampleObject
         It "Creates a PSObject out of JSON input" {