+++ Tests/New-JiraGroup.Tests.ps1
@@ -31,7 +31,7 @@ Describe "New-JiraGroup" {
 
         Mock Invoke-JiraMethod -ModuleName JiraPS -ParameterFilter {$Method -eq 'POST' -and $URI -eq "$jiraServer/rest/api/latest/group"} {
             ShowMockInfo 'Invoke-JiraMethod' 'Method', 'Uri'
-            ConvertFrom-Json2 $testJson
+            ConvertFrom-Json $testJson
         }
 
         # Generic catch-all. This will throw an exception if we forgot to mock something.